rename eventutils.js, fix jsdoc, add filterbox's _getValidFilterNodes() helper
